convert id datatype int to long

+++ ecw/src/main/java/com/ecw/cricketscore/entities/Player.java
@@ -13,15 +13,16 @@ public class Player
 {
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private int id;
+	private Long id;
 	private String name;
 	@OneToMany(mappedBy = "player")
 	private List<TournamentTeamPlayer> tournamentTeamPlayers;
 	
-	public int getId() {
+	
+	public Long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getName() {
